Image upload to S3
As a User I want to upload an image to Amazon S3
As a User I want to preview my image in the browser before uploading it
As a User I want to set a name and a description for my image, including name and description.
As a User I want to see a list with all the images I have uploaded
As a System I should not let the user send an empty image, name or description.
As a System I should not let the user upload a file bigger than 5MB